Add some tests. Commit before trying to change updatePassenger

+++ src/main/java/com/example/demo/domain/Bus.java
@@ -1,6 +1,7 @@
 package com.example.demo.domain;
 
 
+import java.math.BigDecimal;
 import lombok.AllArgsConstructor;
 import lombok.Data;
 import lombok.NoArgsConstructor;
@@ -14,7 +15,7 @@ public class Bus {
   private String busNumber;
   private String registryNumber;
   private Integer amountOfSeats;
-  private Double pricePerKilometre;
+  private BigDecimal pricePerKilometre;
   private Integer routeLength;
 
 }
