+++ src/main/java/com/example/demo/domain/create_requests/CreateBusRequest.java
@@ -1,5 +1,6 @@
 package com.example.demo.domain.create_requests;
 
+import java.math.BigDecimal;
 import lombok.Data;
 
 @Data
@@ -8,7 +9,7 @@ public class CreateBusRequest {
   private String busNumber;
   private String registryNumber;
   private Integer AmountOfSeats;
-  private Double pricePerKilometre;
+  private BigDecimal pricePerKilometre;
   private Integer routeLength;
 
 }
