+++ src/main/java/com/example/demo/domain/create_requests/CreatePassengerRequest.java
@@ -1,11 +1,12 @@
 package com.example.demo.domain.create_requests;
 
+import java.math.BigDecimal;
 import lombok.Data;
 
 @Data
 public class CreatePassengerRequest {
 
   private String idCode;
-  private Double money;
+  private BigDecimal money;
 
 }
